+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+- Add `memoiseWithCycleDetection` and `Cycle`, enabling cycle detection
 - Implement `Task` using `ReaderT`, improving performance
 - Make buildable with GHC 8.0.2 through 8.8.3
 - Switch from the `dependent-map` package to the `dependent-hashmap` for caches